Interfaces for Voting
Added IVoter to represent a voter
Added IVote to represent a vote being conducted
IPlayer implements IVoter to allow players to vote on operation members
The CallVote method of IVote is passed a list of voters, and returns a
Task<bool> representing the outcome of the vote.
The RequestVote method of a voter returns a Task<bool> used to obtain
their vote when they are prepared to give it.