Starting Missions
Missions are now triggered by passing a mission into IGame.StartMission.
The game is then expected to call IPlayer.ChooseOperatives on the
leader. After the leader completes the Task returned by
ChooseOperatives, the OperativesChosen event on IGame is fired.
Modified IGame and IPlayer to handle changes, added code to Game and
relevant tests.
Also altered the Missions and Players properties of IGame to be
IEnumerables, and added tests to assert that order is preserved. This is
vital for games where missions are to be played in a set order.